Casefiles Working

+++ app/controllers/casefiles_controller.rb
@@ -1,14 +1,7 @@
 class CasefilesController < ApplicationController
-  
+
   	def new
-		@casefile = current_user.casefiles.build
-		@casefile.CR="CR test"
-		@casefile.CCN="CCN test"
-		@casefile.CT="CT test"
-		@casefile.CJ="CJ test"
-		@casefile.CA="CA test"
-		@casefile.SAO="SAO test"
-		@casefile.save
+		@casefile = current_user.casefiles.new
 	end
 	
 	def create
@@ -16,12 +9,17 @@ class CasefilesController < ApplicationController
 	end
 	
 	def destroy
-		if @casefile.destroy
-			flash[:casefilesuccess] = "Casefile destroyed"
-			redirect_to :back
-		else
-			flash[:casefileerror] = "Casefile not destroyed"
-			redirect_to :back
+		@casefile = Casefile.find(params[:id])
+		if !@casefile.nil?
+		
+			if @casefile.destroy
+				flash[:casefilesuccess] = "Casefile destroyed"
+				redirect_to :back
+			else
+				flash[:casefileerror] = "Casefile not destroyed"
+				redirect_to :back
+			end
+		
 		end
 	end
